Move things to deriv dir, put GTM&TrackJS to index
